商品削除機能2

+++ app/controllers/items_controller.rb
@@ -38,6 +38,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path unless current_user == @item.user
     if @item.destroy
       flash[:notice] = "削除が完了しました"
       redirect_to root_path
